requirements.txt
beautifulsoup4==4.13.4
chromadb==1.5.0
flask==3.1.2
lxml==5.4.0
matplotlib==3.10.0
openai==1.75.0
python-dotenv==1.1.0
ragas==0.4.3
rank-bm25==0.2.2
rich==14.0.0
sentence-transformers==5.2.2
tqdm==4.67.1
transformers==5.1.0
